Refactor: Split INSERT and UPDATE entry points in Indexer
