+++ benchmarks/src/main/java/io/crate/execution/dml/IndexerBenchmark.java
@@ -100,6 +100,7 @@ public class IndexerBenchmark {
                 table.getReference(ColumnIdent.of("x")),
                 table.getReference(ColumnIdent.of("y"))
             ),
+            null,
             null
         );
         items = IntStream.range(1, 2000).mapToObj(x -> new IndexItem.StaticItem(
