+++ server/src/main/java/io/crate/execution/dml/RawIndexer.java
@@ -105,6 +105,7 @@ public class RawIndexer {
                 txnCtx,
                 nodeCtx,
                 targetRefs,
+                null,
                 returnValues
             );
         });
