+++ server/src/main/java/io/crate/execution/dml/upsert/TransportShardUpsertAction.java
@@ -155,39 +155,17 @@ public class TransportShardUpsertAction extends TransportShardAction<
             }
         }
 
-        UpdateToInsert updateToInsert = null;
-        Indexer indexer;
-        ColumnIdent firstColumnIdent;
-        if (request.updateColumns() != null && request.updateColumns().length > 0) {
-            updateToInsert = new UpdateToInsert(
-                nodeCtx,
-                txnCtx,
-                tableInfo,
-                request.updateColumns(),
-                insertColumns
-            );
-            indexer = new Indexer(
-                partitionValues,
-                tableInfo,
-                indexShard.getVersionCreated(),
-                txnCtx,
-                nodeCtx,
-                updateToInsert.columns(),
-                request.returnValues()
-            );
-            firstColumnIdent = indexer.columns().getFirst().column();
-        } else {
-            indexer = new Indexer(
-                partitionValues,
-                tableInfo,
-                indexShard.getVersionCreated(),
-                txnCtx,
-                nodeCtx,
-                insertColumns,
-                request.returnValues()
-            );
-            firstColumnIdent = indexer.columns().getFirst().column();
-        }
+        Indexer indexer = new Indexer(
+            partitionValues,
+            tableInfo,
+            indexShard.getVersionCreated(),
+            txnCtx,
+            nodeCtx,
+            insertColumns,
+            request.updateColumns(),
+            request.returnValues()
+        );
+        ColumnIdent firstColumnIdent = indexer.columns().getFirst().column();
 
         RawIndexer rawIndexer = null;
         if (firstColumnIdent.equals(SysColumns.RAW)) {
@@ -237,7 +215,6 @@ public class TransportShardUpsertAction extends TransportShardAction<
                     indexShard,
                     tableInfo,
                     partitionValues,
-                    updateToInsert,
                     rawIndexer
                 );
                 if (indexItemResult != null) {
@@ -347,6 +324,7 @@ public class TransportShardUpsertAction extends TransportShardAction<
                 txnCtx,
                 nodeCtx,
                 targetColumns,
+                null,
                 null
             );
         }
@@ -425,7 +403,6 @@ public class TransportShardUpsertAction extends TransportShardAction<
                                       IndexShard indexShard,
                                       DocTableInfo tableInfo,
                                       List<String> partitionValues,
-                                      @Nullable UpdateToInsert updateToInsert,
                                       @Nullable RawIndexer rawIndexer) throws Exception {
         VersionConflictEngineException lastException = null;
         Object[] insertValues = item.insertValues();
@@ -448,7 +425,6 @@ public class TransportShardUpsertAction extends TransportShardAction<
                         // Get most-recent table info, could have changed (new columns, dropped columns)
                         actualTable = schemas.getTableInfo(tableInfo.ident());
                     }
-                    assert updateToInsert != null;
                     assert hasUpdate;
                     String id = item.id();
                     indexItem = PKLookupOperation.withDoc(
@@ -466,7 +442,7 @@ public class TransportShardUpsertAction extends TransportShardAction<
                                 throw new DocumentMissingException(indexShard.shardId(), id);
                             }
                             version.setPlain(doc.getVersion());
-                            return updateToInsert.convert(doc, item.updateAssignments(), insertValues);
+                            return indexer.update(doc, item.updateAssignments(), insertValues);
                         }
                     );
                 }
