+++ server/src/testFixtures/java/io/crate/operation/aggregation/AggregationTestCase.java
@@ -406,7 +406,7 @@ public abstract class AggregationTestCase extends ESTestCase {
             CoordinatorTxnCtx.systemTransactionContext(),
             nodeCtx,
             targetColumns,
-            null
+                null, null
         );
 
         final long startTime = System.nanoTime();
